requirements.txt
mcp[cli]>=1.10.0
lief-extended>=0.17.0
pydantic>=2.0.0
capstone>=5.0.0
keystone-engine>=0.9.2
